+++ .gitignore
@@ -6,7 +6,7 @@
 node_modules
 /.pnp
 .pnp.js
-.package-lock.json
+package-lock.json
 
 # testing
 /coverage
@@ -24,3 +24,6 @@ node_modules
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# s3
+s3.json
