+++ server/models/users.js
@@ -17,6 +17,7 @@ module.exports = (sequelize, DataTypes) => {
       nickname: DataTypes.STRING,
       email: DataTypes.STRING,
       jwt: DataTypes.STRING,
+      profileImageURL: DataTypes.STRING,
     },
     {
       sequelize,
