+++ server/services/multer.js
@@ -1,0 +1,47 @@
+const fs = require("fs");
+const path = require("path");
+const multer = require("multer");
+const multerS3 = require("multer-s3");
+const aws = require("aws-sdk");
+aws.config.loadFromPath(__dirname + "/../config/s3.json");
+
+const s3 = new aws.S3();
+const upload = multer({
+  storage: multerS3({
+    s3: s3,
+    bucket: "images.whatdidyoueat",
+    acl: "public-read-write",
+    key: function (req, file, cb) {
+      cb(null, Date.now() + "." + file.originalname.split(".").pop()); // 이름 설정
+    },
+  }),
+});
+
+// const upload = multer({
+//   // dest: "uploads/",
+//   storage: multer.diskStorage({
+//     destination(req, file, cb) {
+//       cb(null, "uploads/");
+//     },
+//     filename(req, file, cb) {
+//       const ext = path.extname(file.originalname);
+//       cb(null, path.basename(file.originalname, ext) + Date.now() + ext);
+//     },
+//   }),
+//   limits: { fileSize: 5 * 1024 * 1024 },
+// });
+
+// s3.deleteObject(
+//   {
+//     Bucket: "버킷명",
+//     Key: "파일명",
+//   },
+//   function (err, data) {}
+// );
+
+function uploadImg(req, res, next) {}
+
+module.exports = {
+  uploadImg,
+  upload,
+};
